build: 使用cmake进行构建
