+++ .github/workflows/build.yml
@@ -15,8 +15,8 @@ jobs:
       - name: apt install
         run: sudo apt install bison flex libgmp-dev libmpfr-dev libmpc-dev texinfo grub-pc-bin xorriso mtools -y
       - name: install toolchain
-        run: make install-toolchain
+        run: bash meta/cross/install-toolchain.sh
       - name: checkout path
         run: echo $PATH && ls -l $HOME/huangmei-toolchain/bin && ls -l $HOME
       - name: make
-        run: export PATH="$HOME/huangmei-toolchain/bin:$PATH" && make
+        run: export PATH="$HOME/huangmei-toolchain/bin:$PATH" && cmake . -B build && cmake --build build && cmake --build build --target image
