+++ .gitignore
@@ -57,6 +57,7 @@ dkms.conf
 
 # Build
 build
+cmake-build-*
 
 # macOS
 .DS_Store
