+++ kernel/include/log.h
@@ -1,0 +1,6 @@
+#ifndef HUANGMEI_LOG_H
+#define HUANGMEI_LOG_H
+
+void print();
+
+#endif  // HUANGMEI_LOG_H
